fastapi_backend/pyproject.toml
[project]
name = "app"
version = "0.0.6"
description = ""
authors = [{ name = "Anderson Resende", email = "anderson@vinta.com.br" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "fastapi[standard]>=0.115.0,<0.116",
    "asyncpg>=0.29.0,<0.30",
    "fastapi-users[sqlalchemy]>=13.0.0,<14",
    "pydantic-settings>=2.5.2,<3",
    "fastapi-mail>=1.4.1,<2",
    "fastapi-pagination==0.13.3",
    # Schema-only Docling dependency. Provides DoclingDocument, BoundingBox,
    # ProvenanceItem, etc. We do NOT use the full `docling` pipeline (layout
    # models / OCR / TableFormer) — Chandra (via datalab SDK) supplies all
    # OCR + structure, and our adapter maps it to DoclingDocument.
    "docling-core>=2.74,<3",
    "pydantic-ai-slim[openai,google]>=1.80.0,<2",
    "pydantic-graph>=1.80.0,<2",
    "pdf2image>=1.17,<2",
    # Official Datalab (Chandra OCR) SDK — handles auth, polling, retries.
    "datalab-python-sdk>=0.5.0,<1",
    # Async filesystem ops for the local blob backend.
    "aiofiles>=24.1.0,<25",
    # DSPy: typed Signature programming model for LLM/VLM calls. Used by
    # app/refinement/ to keep ARQ-style reasoning checkpoints, traceable
    # I/O, and (later) GEPA optimization on a held-out IoU eval set.
    "dspy-ai>=2.5.0,<3",
    # Pillow: image encoding for VLM payloads + page-render coord checks.
    "pillow>=10.4.0,<12",
]

[dependency-groups]
dev = [
    "pre-commit>=3.4.0,<4",
    "ruff>=0.1.0,<0.2",
    "watchdog>=5.0.3,<6",
    "python-dotenv>=1.0.1,<2",
    "pytest>=8.3.3,<9",
    "pytest-mock>=3.14.0,<4",
    "mypy>=1.13.0,<2",
    "coveralls>=4.0.1,<5",
    "alembic>=1.14.0,<2",
    "pytest-asyncio>=0.24.0,<0.25",
    "mkdocs-material>=9.6.9",
    "mkdocs-material[imaging]>=9.6.9",
    "pytest-cov>=7.1.0",
    "pydantic-evals>=1.80.0",
    "types-aiofiles>=24.1,<25",
]

[tool.uv]
package = false

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.hatch.build.targets.sdist]
include = ["commands"]

[tool.hatch.build.targets.wheel]
include = ["commands"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
